Update launch template and auto scaling group
Associate public IPs in 'network_interfaces' block.

Security groups must be moved into the 'network_interfaces' block if
the block is declared.

'availability_zones' was initially used to point to the AZs
but according to the scaling group resource documentation, it should
only be used when launching instances into the default VPC.

'vpc_zone_identifier' was introduced to reference the AZs.