Remove web-access security group, AZ variables
The security group variable referenced the 'web-access' security group
which is attached to the default VPC. This caused a conflict when
the custom VPC was introduced on 9e90bbb. Therefore, a new security group
was introduced into the template.

The 'availability-zones' variable seemed reduntant since Terraform
doesn't allow you to refer to an entire list using a variable name.
You still have to use index numbers when referencing individual
availability zones.