Add VPC peering connection route
According to the AWS documentation, a route table that directs traffic
to the peering VPC must be added. Otherwise the connection won't work.

I tested the peering connection without adding a route to the route
table in either VPC and was able to ping the EC2 instance in the other
VPC using a private IP.

The routes were added for completeness.